+++ lib/activerecord/null/mimic.rb
@@ -54,6 +54,8 @@ module ActiveRecord
 
       def persisted? = false
 
+      def _read_attribute(_) = nil
+
       def method_missing(method, ...)
         reflections = mimic_model_class.reflect_on_all_associations
         if (reflection = reflections.find { |r| r.name == method })